finish performance/benchmarking tests

+++ perf/index.js
@@ -5,6 +5,7 @@ var path = require('path');
 var mkdirp = require('mkdirp');
 var rimraf = require('rimraf');
 var utils = require('../lib/utils');
+var async = require('async');
 
 var runWriteSpeedBenchmark = require('./write-speed');
 var runReadSpeedBenchmark = require('./read-speed');
@@ -18,25 +19,11 @@ var runUnlinkSpeedBenchmark = require('./unlink-speed');
  * @param {Function} callback
  */
 module.exports = function(options, callback) {
-  runWriteSpeedBenchmark(options, function(err, writeResults) {
-    if (err) {
-      return callback(err);
-    }
-
-    runReadSpeedBenchmark(options, function(err, readResults) {
-      if (err) {
-        return callback(err);
-      }
-
-      runUnlinkSpeedBenchmark(options, function(err, unlinkResults) {
-        if (err) {
-          return callback(err);
-        }
-
-        callback(null, writeResults, readResults, unlinkResults);
-      });
-    });
-  });
+  async.waterfall([
+    runWriteSpeedBenchmark.bind(null, options),
+    runReadSpeedBenchmark.bind(null, options),
+    runUnlinkSpeedBenchmark.bind(null, options)
+  ], callback);
 };
 
 /**
@@ -55,6 +42,11 @@ module.exports.formatResults = function(writeRes, readRes, unlinkRes) {
 
 // NB: If we are running this as a script, go ahead and execute and print out
 if (process.argv[2] === 'exec') {
+  var testsRun = 0;
+  var tests = parseInt(process.argv[3]) || 1;
+  var resultsOut = process.argv[4];
+  var referenceId = utils.createReferenceId().toString('hex');
+
   var TMP_PATH = path.join(os.tmpdir(), 'KFS_PERF_SANDBOX');
   var TABLE_PATH = path.join(TMP_PATH, Date.now().toString());
 
@@ -63,19 +55,42 @@ if (process.argv[2] === 'exec') {
   }
 
   mkdirp.sync(TABLE_PATH);
-  module.exports({
-    tmpPath: TMP_PATH,
-    tablePath: TABLE_PATH
-  }, function(err, writeResults, readResults, unlinkResults) {
-    console.log('Cleaning test environment...')
-    rimraf.sync(TMP_PATH);
 
-    if (err) {
-      return console.error('Error running benchmarks:', err);
-    }
+  var results = [];
 
-    console.info(
-      module.exports.formatResults(writeResults, readResults, unlinkResults)
-    );
-  });
+  function runBenchmarkTests() {
+    module.exports({
+      tmpPath: TMP_PATH,
+      tablePath: TABLE_PATH,
+      referenceId: referenceId
+    }, function(err, writeResults, readResults, unlinkResults) {
+      if (err) {
+        return console.error('Error running benchmarks:', err);
+      }
+
+      testsRun++;
+
+      results.push(module.exports.formatResults(
+        writeResults,
+        readResults,
+        unlinkResults
+      ));
+
+      if (testsRun < tests) {
+        return runBenchmarkTests();
+      }
+
+      console.log('Cleaning test environment...')
+      rimraf.sync(TMP_PATH);
+
+      if (resultsOut) {
+        fs.writeFileSync(resultsOut, JSON.stringify(results));
+        console.info('Results written to %s', resultsOut);
+      } else {
+        console.info(require('util').inspect(results, { depth: null }));
+      }
+    });
+  }
+
+  runBenchmarkTests();
 }
